+++ src/App.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import "./App.css";
 
 import { BrowserRouter as Router } from "react-router-dom";
@@ -6,39 +6,26 @@ import { BrowserRouter as Router } from "react-router-dom";
 import AppAppBar from "./modules/views/AppAppBar";
 import AppFooter from "./modules/views/AppFooter";
 import Routes from "./Router";
-
-import { signout, auth } from "./services/firebase";
+import { Provider, Consumer } from "./AuthContext";
 
 function App() {
-  const [isAuth, setIsAuth] = useState(!!auth.currentUser);
-
-  const setAuthentication = val => {
-    if (!val) {
-      signout();
-      sessionStorage.clear();
-    }
-    setIsAuth(val);
-  };
-
-  auth.onAuthStateChanged(user => {
-    if (user) {
-      setIsAuth(true);
-    } else {
-      setIsAuth(false);
-    }
-  });
-
   return (
-    <Router>
-      <React.Fragment>
-        <AppAppBar 
-        isAuth={isAuth} 
-        setAuthentication={setAuthentication} 
-        />
-        {<Routes isAuth={isAuth} setAuthentication={setAuthentication} />}
-        <AppFooter />
-      </React.Fragment>
-    </Router>
+    <Provider>
+      <Consumer>
+      {({isAuth, setAuthentication}) => (
+         <Router>
+         <React.Fragment>
+           <AppAppBar 
+             isAuth={isAuth} 
+             setAuthentication={setAuthentication} 
+           />
+           {<Routes isAuth={isAuth} setAuthentication={setAuthentication} />}
+           <AppFooter />
+         </React.Fragment>
+       </Router>
+      )}
+      </Consumer>
+    </Provider>
   );
 }
