+++ src/services/firebase.js
@@ -46,13 +46,13 @@ export const adduser = (firstName, lastName, email, typeUser, urlphoto, idusuari
 
 export const addpost = (titulo, autor, lugar, descripcion, fecha, imgLink, idUsuario) => {
   db.collection('Eventos').doc(titulo).set({
-  Titulo: titulo,
-  Autores: autor,
-  Lugar: lugar,
-  Descripcion: descripcion,
-  Fecha: fecha,
-  photoEvent: imgLink
-});
+    Titulo: titulo,
+    Autores: autor,
+    Lugar: lugar,
+    Descripcion: descripcion,
+    Fecha: fecha,
+    photoEvent: imgLink
+  });
 };
